src/services/products/requirements.txt
python-dotenv
fastapi[all]
requests
google-cloud-storage
google-cloud-firestore
pytest